fixation repositories

+++ src/main/java/tn/esprit/gestionprojet/repositories/EquipeRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.gestionprojet.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.gestionprojet.entities.Equipe;
 
-public interface EquipeRepository extends JpaRepository<Equipe, Integer> {
+public interface EquipeRepository extends JpaRepository<Equipe, Long> {
 }
